SDC split into dllmain, configuration, and cartridge
inital split of dll.cpp into three parts
Added some comments in libcommon and becker
Changed '...' to '>' on MPI config
