+++ sdc/sdc.vcxproj
@@ -96,11 +96,14 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="sdc.cpp" />
+    <ClCompile Include="sdc_dllmain.cpp" />
+    <ClCompile Include="sdc_cartridge.cpp" />
+    <ClCompile Include="sdc_configuration.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="sdc.h" />
     <ClInclude Include="resource.h" />
+    <ClInclude Include="sdc_cartridge.h" />
+    <ClInclude Include="sdc_configuration.h" />
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="sdc.rc" />
